refactor(bim): extract the placeable-file scene machinery from FilesSection
Models and Files both need the registry sync, seeding, click-to-place and
gizmo edit, so it moves to a hook rather than being duplicated.
